+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 ### Added
 
-- Added "Else-If" (or "ElseIf") to "If" statement
+- Added "Else-If" (or "ElseIf") to the "If" statement
+- Added "If ... =>/⇒" guard format to the "If" statement
 - Added "GetFileInfo()" to retrieve information on one or more file/directory paths.
   Variable length arguments and supports polymorphism with lists and dictionaries.
 - Added "Head()" and "Tail()" to return a subset of lists.